rl/requirements.txt
# GRPO Training Requirements
# Install with: pip install -r requirements.txt

# Core ML
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
trl>=0.7.4

# Chess
python-chess>=1.10.0
stockfish>=3.28.0

# Training utilities
unsloth @ git+https://github.com/unslothai/unsloth.git
bitsandbytes>=0.41.0
datasets>=2.14.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging
wandb>=0.16.0
tqdm>=4.66.0
